move datetime parser to core project
